Move BigQueryRunConfig to a separate file
