+++ metaphor/bigquery/extractor.py
@@ -1,11 +1,8 @@
 import json
 import logging
 import re
-from dataclasses import dataclass
-from dataclasses import field as dataclass_field
-from typing import Any, List, Mapping, Optional, Sequence, Union
+from typing import Any, List, Mapping, Sequence, Union
 
-from serde import deserialize
 from smart_open import open
 
 try:
@@ -28,24 +25,12 @@ from metaphor.models.metadata_change_event import (
     SQLSchema,
 )
 
+from metaphor.bigquery.config import BigQueryRunConfig
 from metaphor.common.event_util import EventUtil
-from metaphor.common.extractor import BaseExtractor, RunConfig
+from metaphor.common.extractor import BaseExtractor
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
-
-
-@deserialize
-@dataclass
-class BigQueryRunConfig(RunConfig):
-    # Path to service account's JSON key file
-    key_path: str
-
-    # Project ID to use. Use the service account's default project if not set
-    project_id: Optional[str] = None
-
-    # Filters for dataset names (any match will be included)
-    dataset_filters: List[str] = dataclass_field(default_factory=lambda: [r".*"])
 
 
 def build_client(config: BigQueryRunConfig):
